[boo] move gemm under the new op registration mechanism
This makes it possible to use unified scripts for benchmarking,
numerics, etc.

Also add a clause in the cache prepopulator test to include gemms. Note
that the compilation is currently broken on CPU, unsure if #998 is
related.
